Refactor version engine storage layer
