+++ backend/src/version_engine/storage/__init__.py
@@ -1,0 +1,25 @@
+"""L6 storage substrate for Version Engine."""
+
+from src.version_engine.storage.io_strategy import (
+    IOStorageStrategy,
+    ObjectWriteLayout,
+    ObjectWritePlan,
+    ObjectWriteRoute,
+)
+from src.version_engine.storage.object_store import (
+    FileSystemBackend,
+    ObjectStore,
+    StorageBackend,
+    stage_object_writes,
+)
+
+__all__ = [
+    "FileSystemBackend",
+    "IOStorageStrategy",
+    "ObjectStore",
+    "ObjectWriteLayout",
+    "ObjectWritePlan",
+    "ObjectWriteRoute",
+    "StorageBackend",
+    "stage_object_writes",
+]
