+++ backend/src/version_engine/storage/backends/__init__.py
@@ -1,0 +1,17 @@
+"""Storage backend implementations for Version Engine L6."""
+
+from src.version_engine.storage.backends.s3 import (
+    CachedStorageBackend,
+    ObjectLocation,
+    ObjectStorageLayout,
+    ObjectWriteBatch,
+    S3StorageBackend,
+)
+
+__all__ = [
+    "CachedStorageBackend",
+    "ObjectLocation",
+    "ObjectStorageLayout",
+    "ObjectWriteBatch",
+    "S3StorageBackend",
+]
