+++ backend/src/version_engine/write_engine/engine.py
@@ -16,7 +16,7 @@ import time
 from datetime import datetime, timezone
 from typing import Callable
 
-from src.version_engine.write_engine.object_store import (
+from src.version_engine.storage.object_store import (
     ObjectStore,
     stage_object_writes,
 )
