+++ backend/tests/version_engine/conftest.py
@@ -14,7 +14,7 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from src.version_engine.write_engine.object_store import ObjectStore
+from src.version_engine.storage.object_store import ObjectStore
 from src.version_engine.infrastructure.supabase.repo_manager import VersionRepoManager
