+++ backend/tests/version_engine/test_git_native_parity_matrix.py
@@ -12,7 +12,7 @@ from unittest.mock import MagicMock
 
 import pytest
 from src.version_engine.write_engine.merge import merge_file_sets, three_way_merge
-from src.version_engine.write_engine.object_store import ObjectStore
+from src.version_engine.storage.object_store import ObjectStore
 from src.version_engine.write_engine.path_utils import normalize_path
 from src.version_engine.infrastructure.supabase.scope_manager import ScopeManager
