+++ backend/tests/version_engine/test_server_repo.py
@@ -14,7 +14,7 @@ import threading
 
 import pytest
 
-from src.version_engine.write_engine.object_store import ObjectStore
+from src.version_engine.storage.object_store import ObjectStore
 from src.version_engine.write_engine.git_object_format import MODE_DIR, MODE_FILE, TreeEntry, encode_tree
